:bento: (lld): onboarding success
