+++ apps/ledger-live-desktop/src/renderer/components/SyncOnboarding/Manual/EuropaCompletionView.tsx
@@ -1,0 +1,2 @@
+import StaxCompletionView from "./StaxCompletionView";
+export default StaxCompletionView;
